+++ src/RestApi/Emulator/IEmulatorFilter.ts
@@ -1,0 +1,4 @@
+
+export interface IEmulatorFilter {
+	organizationUid?: string;
+}