+++ tests/Integration/RestAPI/Emulator/RestAPI.spec.ts
@@ -32,6 +32,16 @@ describe('RestAPI - Emulator', () => {
 		should(emulator.createdAt instanceof Date).be.true();
 	});
 
+	it('should get list of emulators of organization', async () => {
+		const [emulator] = await api.emulator.list({ organizationUid: opts.organizationUid! });
+		should(emulator instanceof Emulator).be.true();
+		should(emulator.uid).be.not.empty();
+		emulatorUid = emulator.uid;
+		should(emulator.duid).be.not.empty();
+		should(emulator.name).be.not.empty();
+		should(emulator.createdAt instanceof Date).be.true();
+	});
+
 	it('should delete old emulator', async () => {
 		await api.emulator.delete(emulatorUid);
 		should(true).be.true();